LayoutTests/imported/w3c/web-platform-tests/storage/estimate-usage-details-indexeddb.https.tentative.any.worker-expected.txt


FAIL estimate() resolves to dictionary with usageDetails member assert_equals: expected "object" but got "undefined"
FAIL estimate() usage details reflects increase in indexedDB after large value is stored promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'estimate.usageDetails.indexedDB')"
